refactor(gui): extract worker, controller, widgets and app modules (phases 2-4) (#48)
QualityScaler.py shrinks from 1820 lines to a 12-line launcher shim; the GUI
now lives in qualityscaler/gui/ with the toolkit layer (customtkinter)
confined to widgets.py and app.py. The pipeline subprocess entry point moves
to gui/worker.py so spawned children no longer import the whole GUI, and the
toolkit-free controller (validation, settings building, process control) is
now unit tested headlessly.

Part of #44
