+++ src/qualityscaler/gui/__init__.py
@@ -1,6 +1,8 @@
-"""Toolkit-free GUI support modules for QualityScaler.
+"""GUI package for QualityScaler.
 
-Modules in this package must stay importable without customtkinter so they
-can be exercised by the headless unit-test environment. ``media_info`` is the
-one exception: it needs cv2/numpy and must not be imported from here.
+The toolkit layer (customtkinter/tkinter) is confined to ``widgets`` and
+``app``; every other module must stay importable without customtkinter so it
+can be exercised by the headless unit-test environment (``media_info`` and
+``widgets`` additionally need cv2/numpy). For that reason this package does
+not import anything eagerly — use ``qualityscaler.gui.app.main`` to launch.
 """
